Add a post-parse, pre-render hook.
This implements the ability for the user to pass in a func that
will be called on the ctx after imports, links, etc. have been resolved.
This was necessary for the gui editor to get the real title out of the
ast without having to parse the rendered html.

Currently this is only usable from the internal API, which means
gdndc is able to do stuff that an independent implementation would
not be able to do. I don't like this, but I don't want to just
jam the context into the public API.